Remove unneeded complexity
