--- pkg/metrics/taggable.go
@@ -1,100 +1,0 @@
-// Copyright (c) 2021-2024 SigScalr, Inc.
-//
-// This file is part of SigLens Observability Solution
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package metrics
-
-import (
-	"fmt"
-	"sort"
-	"strings"
-)
-
-type Taggable interface {
-	GetValue(key string) (string, bool)
-	SetValue(key string, value string)
-	RemoveTag(key string)
-	HasTag(key string) bool
-	SetTags(tags map[string]string)
-	Id() string
-}
-
-// Implements the Taggable interface for a Prometheus metric.
-type TaggedMetric struct {
-	tags map[string]string
-}
-
-func NewTaggedMetric() *TaggedMetric {
-	return &TaggedMetric{
-		tags: make(map[string]string),
-	}
-}
-
-func (t *TaggedMetric) GetValue(key string) (string, bool) {
-	value, exists := t.tags[key]
-	return value, exists
-}
-
-func (t *TaggedMetric) SetValue(key string, value string) {
-	if t.tags == nil {
-		t.tags = make(map[string]string)
-	}
-	t.tags[key] = value
-}
-
-func (t *TaggedMetric) RemoveTag(key string) {
-	delete(t.tags, key)
-}
-
-func (t *TaggedMetric) HasTag(key string) bool {
-	_, exists := t.tags[key]
-	return exists
-}
-
-func (t *TaggedMetric) SetTags(tags map[string]string) {
-	t.tags = make(map[string]string)
-	for k, v := range tags {
-		t.tags[k] = v
-	}
-}
-
-func (t *TaggedMetric) Id() string {
-	metricName, ok := t.tags["__name__"] // For Prometheus metrics.
-	if !ok {
-		metricName = ""
-	}
-
-	sortedKeys := make([]string, 0, len(t.tags))
-	for k := range t.tags {
-		if k != "__name__" {
-			sortedKeys = append(sortedKeys, k)
-		}
-	}
-
-	sort.Strings(sortedKeys)
-
-	var id strings.Builder
-	id.WriteString(metricName + "{")
-	for i, key := range sortedKeys {
-		if i > 0 {
-			id.WriteString(",")
-		}
-		id.WriteString(fmt.Sprintf("%s=%s", key, t.tags[key]))
-	}
-	id.WriteString("}")
-
-	return id.String()
-}
